Rename CustomEquipmentSlot to CustomVisualEquipType, adjust comments
