+++ ExampleMod/Common/CustomVisualEquipType/AnglerEarring_Earrings.png
