+++ ExampleMod/Common/CustomVisualEquipType/EarringsDrawLayer.cs
@@ -3,12 +3,12 @@ using Terraria;
 using Terraria.ModLoader;
 using Microsoft.Xna.Framework;
 
-namespace ExampleMod.Common.CustomEquipmentSlot
+namespace ExampleMod.Common.CustomVisualEquipType
 {
-	// Note: To fully understand this example, please start by reading https://github.com/tModLoader/tModLoader/tree/stable/ExampleMod/Common/CustomEquipmentSlot/README.md
+	// Note: To fully understand this example, please start by reading https://github.com/tModLoader/tModLoader/tree/stable/ExampleMod/Common/CustomVisualEquipType/README.md
 
 	/// <summary>
-	/// EarringsDrawLayer is a PlayerDrawLayer that renders a custom equipment slot. 
+	/// EarringsDrawLayer is a PlayerDrawLayer that renders a custom EquipType. 
 	/// </summary>
 	public class EarringsDrawLayer : PlayerDrawLayer
 	{
@@ -33,7 +33,7 @@ namespace ExampleMod.Common.CustomEquipmentSlot
 			) + drawPlayer.headPosition + drawInfo.headVect + Main.OffsetsPlayerHeadgear[drawPlayer.bodyFrame.Y / drawPlayer.bodyFrame.Height] * drawPlayer.gravDir;
 
 			// Source rectangle. Player.bodyFrame corresponds to the animation frame (pose) that the player is currently in.
-			// Vanilla player layers have a separate image for each pose to account for how layers move in relation to each other while the player is animating, but this example doesn't need that so we use the same image all the time.
+			// Vanilla player layers for EquipTypes have a separate image for each pose to account for how layers move in relation to each other while the player is animating, but this example doesn't need that so we use the same image all the time.
 			Rectangle drawFrame = drawPlayer.bodyFrame;
 			drawFrame.Y = 0;
