+++ ExampleMod/Common/CustomVisualEquipType/EarringsLoader.cs
@@ -5,19 +5,19 @@ using System.Collections.Generic;
 using Terraria.ID;
 using Terraria.ModLoader;
 
-namespace ExampleMod.Common.CustomEquipmentSlot
+namespace ExampleMod.Common.CustomVisualEquipType
 {
-	// Note: To fully understand this example, please start by reading https://github.com/tModLoader/tModLoader/tree/stable/ExampleMod/Common/CustomEquipmentSlot/README.md
+	// Note: To fully understand this example, please start by reading https://github.com/tModLoader/tModLoader/tree/stable/ExampleMod/Common/CustomVisualEquipType/README.md
 
 	/// <summary>
-	/// EarringsLoader manages loading earring equipment slot textures. 
+	/// EarringsLoader manages loading earring equipment textures for each item using this custom EquipType. 
 	/// </summary>
 	public class EarringsLoader : ModSystem {
-		/// <summary> Maps an item type to it's corresponding earring equipment slot texture. </summary>
+		/// <summary> Maps an item type to it's corresponding earring equipment texture. </summary>
 		internal static readonly Dictionary<int, Asset<Texture2D>> earringItemToTexture = new();
 
 		public override void ResizeArrays() {
-			// After all content is loaded, autoload earring equipment slot textures
+			// After all content is loaded, autoload earring equipment textures
 			for (int i = ItemID.Count; i < ItemLoader.ItemCount; i++) {
 				ModItem modItem = ItemLoader.GetItem(i);
 				if (modItem.GetType().GetAttribute<AutoloadEquip_EarringAttribute>() != null) {
@@ -25,8 +25,8 @@ namespace ExampleMod.Common.CustomEquipmentSlot
 				}
 			}
 
-			// Add an earring equipment slot to a vanilla item.
-			AddEarringTexture(ItemID.AnglerEarring, "ExampleMod/Common/CustomEquipmentSlot/AnglerEarring_Earrings");
+			// Add an earring equipment texture to a vanilla item.
+			AddEarringTexture(ItemID.AnglerEarring, "ExampleMod/Common/CustomVisualEquipType/AnglerEarring_Earrings");
 		}
 
 		public static void AddEarringTexture(int type, string texture) {
