+++ ExampleMod/Common/CustomVisualEquipType/RubyEarrings.cs
@@ -2,9 +2,9 @@ using Terraria.Enums;
 using Terraria;
 using Terraria.ModLoader;
 
-namespace ExampleMod.Common.CustomEquipmentSlot
+namespace ExampleMod.Common.CustomVisualEquipType
 {
-	// A vanity accessory showcasing our custom earring equipment slot
+	// A vanity accessory showcasing our custom earring EquipType
 	[AutoloadEquip_Earring]
 	public class RubyEarrings : ModItem
 	{
