+++ ExampleMod/Common/CustomVisualEquipType/RubyEarrings.png
