+++ ExampleMod/Common/CustomVisualEquipType/RubyEarrings_Earrings.png
