+++ ExampleMod/Common/CustomVisualEquipType/SapphireEarrings.png
