+++ ExampleMod/Common/CustomVisualEquipType/SapphireEarrings_Earrings.png
