+++ ExampleMod/Content/Items/Accessories/ExampleShield.cs
@@ -6,9 +6,13 @@ using Terraria.ModLoader;
 
 namespace ExampleMod.Content.Items.Accessories
 {
-	[AutoloadEquip(EquipType.Shield)] // Load the spritesheet you create as a shield for the player when it is equipped.
+	[AutoloadEquip(EquipType.Shield)]
 	public class ExampleShield : ModItem
 	{
+		public override void Load() { 
+			//EquipLoader.AddEquipTexture(Mod, $"{Texture}_{EquipType.Shield}", EquipType.Shield, this);
+		}
+
 		public override void SetDefaults() {
 			Item.width = 24;
 			Item.height = 28;
